refactor: hooks for Draggable TextBox
