+++ frontend/src/components/DraggableTextBox.tsx
@@ -1,88 +1,16 @@
 import type React from 'react'
 import { EditorContent } from '@tiptap/react'
 import styles from './DraggableTextBox.module.css'
-import { useState, useCallback, useEffect } from 'react'
+import { useEffect, useCallback } from 'react'
+import { useDrag } from '@/hooks/useDrag'
+import { useResize } from '@/hooks/useResize'
 import type { TextBox } from '@/types/Slide'
+import type { ResizeDivs } from '@/types/DraggableTextBox'
 
 type Props = {
   textbox: TextBox
 }
 
-import type {
-  Direction,
-  ResizeOptions,
-  ResizeDivs,
-} from '@/types/DraggableTextBox'
-
-const handleResize = (
-  e: MouseEvent,
-  resizeStart: { x: number; y: number },
-  size: { width: number; height: number },
-  position: { x: number; y: number },
-  direction: Direction,
-): ResizeOptions => {
-  const deltaX = e.clientX - resizeStart.x
-  const deltaY = e.clientY - resizeStart.y
-
-  const options: { [key in Direction]: ResizeOptions } = {
-    north: {
-      width: size.width,
-      height: size.height - deltaY,
-      x: position.x,
-      y: position.y + deltaY,
-    },
-    northEast: {
-      width: size.width + deltaX,
-      height: size.height - deltaY,
-      x: position.x,
-      y: position.y + deltaY,
-    },
-    east: {
-      width: size.width + deltaX,
-      height: size.height,
-      x: position.x,
-      y: position.y,
-    },
-    southEast: {
-      width: size.width + deltaX,
-      height: size.height + deltaY,
-      x: position.x,
-      y: position.y,
-    },
-    south: {
-      width: size.width,
-      height: size.height + deltaY,
-      x: position.x,
-      y: position.y,
-    },
-    southWest: {
-      width: size.width - deltaX,
-      height: size.height + deltaY,
-      x: position.x + deltaX,
-      y: position.y,
-    },
-    west: {
-      width: size.width - deltaX,
-      height: size.height,
-      x: position.x + deltaX,
-      y: position.y,
-    },
-    northWest: {
-      width: size.width - deltaX,
-      height: size.height - deltaY,
-      x: position.x + deltaX,
-      y: position.y + deltaY,
-    },
-    default: {
-      width: size.width,
-      height: size.height,
-      x: position.x,
-      y: position.y,
-    },
-  }
-
-  return options[direction]
-}
 const handleResizeDivs: ResizeDivs[] = [
   { direction: 'north', className: styles.resizeHandleNorth },
   { direction: 'northEast', className: styles.resizeHandleNorthEast },
@@ -95,96 +23,31 @@ const handleResizeDivs: ResizeDivs[] = [
 ]
 
 const DraggableTextBox: React.FC<Props> = ({ textbox }) => {
-  const [isDragging, setIsDragging] = useState(false)
-  const [dragStart, setDragStart] = useState<{ x: number; y: number } | null>(
-    null,
-  )
-  const [position, setPosition] = useState({ x: textbox.x, y: textbox.y })
-  const [size, setSize] = useState({
-    width: textbox.width,
-    height: textbox.height,
-  })
-  const [isResizing, setIsResizing] = useState(false)
-  const [resizeDirection, setResizeDirection] = useState<Direction>('default')
-  const [resizeStart, setResizeStart] = useState<{ x: number; y: number }>({
-    x: size.width,
-    y: size.height,
-  })
+  const {
+    position: dragPosition,
+    handleDragMouseDown,
+    handleDragMouseMove,
+    handleDragMouseUp,
+  } = useDrag({ x: textbox.x, y: textbox.y })
+
+  const {
+    isResizing,
+    size,
+    position: resizePosition,
+    handleResizeMouseDown,
+    handleResizeMouseMove,
+    handleResizeMouseUp,
+  } = useResize({ width: textbox.width, height: textbox.height }, dragPosition)
 
   const style: React.CSSProperties = {
-    transform: `translate3d(${position.x}px, ${position.y}px, 0)`,
+    transform: `translate3d(${resizePosition.x}px, ${resizePosition.y}px, 0)`,
     position: 'absolute',
     width: `${size.width}px`,
     height: `${size.height}px`,
     boxSizing: 'border-box',
     outline: textbox.isSelected ? 'solid 1px blue' : 'none',
-    userSelect: 'none', // Prevent text selection(入力の無効化はtiptapにメソッドが存在する為、注意が必要)
+    userSelect: 'none',
   }
-
-  const handleDragMouseDown = useCallback(
-    (e: React.MouseEvent) => {
-      textbox.isSelected = true
-      setDragStart({ x: e.clientX - position.x, y: e.clientY - position.y })
-      setIsDragging(true)
-    },
-    [position.x, position.y, textbox],
-  )
-
-  const handleDragMouseMove = useCallback(
-    (e: MouseEvent) => {
-      if (!isDragging || !dragStart) return
-      setPosition({
-        x: e.clientX - dragStart.x,
-        y: e.clientY - dragStart.y,
-      })
-    },
-    [isDragging, dragStart],
-  )
-
-  const handleDragMouseUp = useCallback(() => {
-    if (!isDragging) return
-    setIsDragging(false)
-    setDragStart(null)
-    textbox.isSelected = false
-  }, [isDragging, textbox])
-
-  const handleResizeMouseDown = useCallback(
-    (e: React.MouseEvent, direction: Direction) => {
-      e.stopPropagation()
-      setResizeDirection(direction)
-      setIsResizing(true)
-      setResizeStart({ x: e.clientX, y: e.clientY })
-      textbox.isSelected = true
-    },
-    [textbox],
-  )
-  const handleResizeMouseMove = useCallback(
-    (e: MouseEvent) => {
-      if (!isResizing) return
-      const newOptions = handleResize(
-        e,
-        resizeStart,
-        size,
-        position,
-        resizeDirection,
-      )
-      if (newOptions.width < -10 || newOptions.height < -10) return
-
-      setPosition({ x: newOptions.x, y: newOptions.y })
-      setSize({
-        width: newOptions.width,
-        height: newOptions.height,
-      })
-      setResizeStart({ x: e.clientX, y: e.clientY })
-    },
-    [isResizing, resizeStart, size, position, resizeDirection],
-  )
-  const handleResizeMouseUp = useCallback(() => {
-    if (!isResizing) return
-    setIsResizing(false)
-    setResizeDirection('default')
-    textbox.isSelected = false
-  }, [isResizing, textbox])
 
   const handleMouseMove = useCallback(
     (e: MouseEvent) => {
@@ -193,6 +56,7 @@ const DraggableTextBox: React.FC<Props> = ({ textbox }) => {
     },
     [handleDragMouseMove, handleResizeMouseMove],
   )
+
   const handleMouseUp = useCallback(() => {
     handleDragMouseUp()
     handleResizeMouseUp()
@@ -205,14 +69,7 @@ const DraggableTextBox: React.FC<Props> = ({ textbox }) => {
       document.removeEventListener('mousemove', handleMouseMove)
       document.removeEventListener('mouseup', handleMouseUp)
     }
-  }, [
-    handleDragMouseMove,
-    handleDragMouseUp,
-    handleMouseMove,
-    handleMouseUp,
-    handleResizeMouseMove,
-    handleResizeMouseUp,
-  ])
+  }, [handleMouseMove, handleMouseUp])
 
   return (
     <div
